Move data files to a subdirectory, so they won't be formulas.
