+++ dartium.rb
@@ -3,8 +3,8 @@ require 'formula'
 class Dartium < Formula
   homepage "https://www.dartlang.org"
 
-  require_relative 'dev_info'
-  require_relative 'stable_info'
+  require_relative 'data/dev_info'
+  require_relative 'data/stable_info'
 
   version DartStable::VERSION
   url "https://storage.googleapis.com/dart-archive/#{DartStable::DARTIUM_FILE}"
